+++ mobgap/orientation_estimation/__init__.py
@@ -2,4 +2,4 @@
 
 __all__ = ["MadgwickAHRS"]
 
-from mobgap._gaitmap.trajectory_reconstruction.orientation_methods._madgwick import MadgwickAHRS
+from mobgap.orientation_estimation._madgwick import MadgwickAHRS